article_service/requirements.txt
Django==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
newspaper3k==0.2.8
spacy==3.6.0
tldextract==5.1.1
mongoengine==0.27.0
django-rest-framework-mongoengine==3.4.1